requirements.txt
flask
firebase_admin
easyocr
google.generativeai